+++ src/components/game/reducers/defeat.test.ts
@@ -1,6 +1,8 @@
 import { defeatReducer } from './defeat'
 import { CellStateStage, CellState } from '../../../common/game-types'
 import { playingGameState, lostGameState } from '../../../__mocks__/game-states'
+import storage from '../../../common/storage'
+import { sequenceFillData } from '../scoring'
 
 describe('defeatReducer is called in repetition', () => {
   test('should end with all mines .clicked (CellStateStage.TESTED)', () => {
@@ -25,6 +27,47 @@ describe('defeatReducer is called in repetition', () => {
       .filter(cell => cell.stage === CellStateStage.TESTED && cell.fill > 8)
 
     expect(clickedMines.length).toBe(allMines.length)
+  })
+
+  test('should animate in about 13 batches', () => {
+    let
+      times = 0,
+      releasedMines: number,
+      gameState = { ...playingGameState }
+    const [board, config] = sequenceFillData('kk4Aw2iMaS1styvRK7kgiz32pmqWi+BK+GhySxaZp9ljNV1j5tTDdJ2rnf1HB36F0eLpyFggA')
+    storage.config = { ...storage.config, ...config }
+    const findReleasedMines = () => gameState.board
+      .flat()
+      .filter(cell => cell.stage === CellStateStage.RELEASED && cell.fill > 8)
+      .length
+    gameState.board = board
+    gameState.board.forEach(
+      (r: CellState[]) => r.forEach(
+        (c: CellState) => c.stage = CellStateStage.RELEASED
+      )
+    )
+
+    const initialReleasedMines = findReleasedMines()
+    do {
+      gameState = defeatReducer(gameState)
+      releasedMines = findReleasedMines()
+      times++
+    } while (releasedMines > 0 )
+
+    expect(initialReleasedMines).toBeGreaterThan(releasedMines)
+    expect(releasedMines).toBe(0)
+
+    expect(times).toBeGreaterThan(11)
+    expect(times).toBeLessThan(15)
+
+    const bursts = gameState.board
+      .flat()
+      .filter(cell => cell.burst)
+      .length
+
+    // bursts includes one more: the original misclick
+    // times includes one more: the last iteration does nothing
+    expect(bursts).toBe(times)
   })
 
   test('should return unchanged state when no untouched mines were found', () => {